extra: split ci_inspect_logs.py into collect + generate scripts
Refactor the monolithic ci_inspect_logs.py into three focused scripts:

- ci_shared.py: shared constants (SKIP/PASS/WARNING/EXPECTED_ERROR/ERROR/
  FAILURE), display icon arrays, and data classes (LoaderEntry, TestEntry,
  TestGroup) with to_dict()/from_dict() for JSON serialization.

- ci_collect_data.py: reads environment variables and CI artifact files,
  builds all data structures, and serializes the result to a JSON file.
  Groups are not exported but can be reconstructed by calling track() on
  each test entry.
  Usage: ci_collect_data.py <json_file>

- ci_generate_report.py: reads the JSON file produced by ci_collect_data.py,
  reconstructs all data structures, and writes the same three output files
  (results, summary, full report) as the original script.
  Usage: ci_generate_report.py <json_file> <results> <summary> <full_report>

Update package_core.yml to invoke the two scripts in sequence.
