+++ .github/workflows/package_core.yml
@@ -111,7 +111,7 @@ jobs:
   # for later jobs.
 
   build-board:
-    # NOTE: this name is hardcoded in ci_inspect_logs.py
+    # NOTE: this name is hardcoded in ci_collect_data.py
     name: Build for ${{ matrix.board }}
     runs-on: ubuntu-latest
     needs:
@@ -310,7 +310,7 @@ jobs:
   # This step never fails, so that expected errors do not block the CI run.
 
   test-core:
-    # NOTE: this name is hardcoded in ci_inspect_logs.py
+    # NOTE: this name is hardcoded in ci_collect_data.py
     name: Test ${{ matrix.board }}:${{ matrix.link_mode }}
     runs-on: ubuntu-latest
     needs:
@@ -395,7 +395,7 @@ jobs:
 
 
   # Collect all build and test reports, analyze them and produce a summary.
-  # See ci_inspect_logs.py for details.
+  # See ci_collect_data.py and ci_generate_report.py for details.
 
   verify-core:
     runs-on: ubuntu-latest
@@ -429,8 +429,9 @@ jobs:
           export WORKFLOW_JOBS=$(gh run view ${{ github.run_id }} --attempt ${{ github.run_attempt }} --json jobs --jq '.jobs')
           # Archive the input set of reports for later inspection
           tar jchf build-reports-${{ needs.build-env.outputs.CORE_HASH }}.tar.bz2 zephyr-* arduino-*.json
-          # Run the log inspection script
-          extra/ci_inspect_logs.py result summary full_log
+          # Collect all CI data into a JSON file, then generate the report
+          extra/ci_collect_data.py build_info.json
+          extra/ci_generate_report.py build_info.json result summary full_log
           cat result
 
           # Display the summary and full log in the step summary
