+++ extra/ci_collect_data.py
@@ -1,0 +1,206 @@
+#!/usr/bin/env python3
+
+# Copyright (c) Arduino s.r.l. and/or its affiliated companies
+# SPDX-License-Identifier: Apache-2.0
+
+# Script to analyze CI build and test logs and collect results into a JSON file.
+#
+# The script reads build artifacts and test reports generated during a GitHub
+# Actions CI run for Arduino Zephyr cores. It aggregates results by board,
+# artifact, and sketch, and serializes all data to a JSON file for further
+# processing by ci_generate_report.py.
+#
+# This script expects the following inputs:
+#
+#  - Environment variables:
+#     - ALL_BOARD_DATA: JSON array of all board metadata, generated by get_board_details.sh
+#     - WORKFLOW_JOBS: JSON array of job names and URLs, generated by workflow YAML
+#     - GITHUB_REPOSITORY, GITHUB_RUN_ID: workflow context variables for writing links
+#
+#  - Files in the current directory:
+#     - zephyr-<variant>.config: Kconfig settings, generated at loader build time
+#     - zephyr-<variant>.warnings: Loader compilation warnings
+#     - zephyr-<variant>.meminfo: Loader memory usage report
+#     - arduino-<subarch>-<board>-link_mode=<mode>.json: Compilation test report files
+#     - variants/<variant>/known_example_issues.txt: Expected errors for that variant
+#
+# This script produces one output file:
+#
+#  - <json_file>: JSON file with all collected data, to be consumed by ci_generate_report.py
+#
+# Usage: ci_collect_data.py [<json_file>]
+#  If no argument is given, output is written to /dev/stdout.
+
+from collections import defaultdict
+import json
+import os
+import re
+import sys
+
+import ci_shared
+from ci_shared import (
+    SKIP, PASS, WARNING, EXPECTED_ERROR, ERROR, FAILURE,
+    LoaderEntry, TestEntry, TestGroup,
+)
+
+# Global Data Structures
+# ----------------------
+
+# Loader build results, one per board
+BOARD_LOADERS = {}                                          # { board: LoaderEntry() }
+
+ALL_TESTS = []                                               # [ TestEntry() ]
+
+# Test results grouped by artifact, board, and artifact/sketch
+# (grouping sketch results across different artifacts is really confusing)
+ARTIFACT_TESTS = defaultdict(TestGroup)                     # { artifact: TestGroup() }
+BOARD_TESTS = defaultdict(TestGroup)                        # { board: TestGroup() }
+SKETCH_TESTS = defaultdict(lambda: defaultdict(TestGroup))  # { artifact: { sketch: TestGroup() } }
+
+def log_test(artifact, board, sketch, link_mode, exceptions, status, issues, job_link=None):
+    """
+    Logs individual test results into the global test tracking structures.
+    """
+
+    # Ensure issues is a list
+    if isinstance(issues, str):
+        issues = [ issues ]
+
+    # Create the test entry
+    excepted = any(pattern.match(sketch) for pattern in exceptions)
+    status = EXPECTED_ERROR if excepted and status == ERROR else status
+    test_entry = TestEntry(artifact, board, sketch, link_mode, excepted, status, issues, job_link)
+
+    # Track in global structures
+    ARTIFACT_TESTS[artifact].track(test_entry)
+    BOARD_TESTS[board].track(test_entry)
+    SKETCH_TESTS[artifact][sketch].track(test_entry)
+    ALL_TESTS.append(test_entry)
+
+# ---------------------------------------------------------------------------
+# Main Logic
+# ---------------------------------------------------------------------------
+
+# Environment Variable Checks
+ALL_BOARD_DATA_STR = os.environ.get('ALL_BOARD_DATA')
+WORKFLOW_JOBS_STR = os.environ.get('WORKFLOW_JOBS')
+GITHUB_REPOSITORY = os.environ.get('GITHUB_REPOSITORY')
+GITHUB_RUN_ID = os.environ.get('GITHUB_RUN_ID')
+JOB_LINK_STEM = f"https://github.com/{GITHUB_REPOSITORY}/actions/runs/{GITHUB_RUN_ID}"
+
+if not ALL_BOARD_DATA_STR or not GITHUB_REPOSITORY or not GITHUB_RUN_ID:
+    print("Not in a Github CI run, cannot proceed.")
+    sys.exit(0)
+
+if not len(sys.argv) in (1, 2):
+    print("Usage: ci_collect_data.py [<json_file>]")
+    sys.exit(1)
+
+json_file = sys.argv[1] if len(sys.argv) == 2 else "/dev/stdout"
+
+# Load board data and job URLs from environment variables
+ALL_BOARD_DATA = json.loads(ALL_BOARD_DATA_STR)
+ALL_BOARD_DATA = { b['board']: b for b in ALL_BOARD_DATA }
+
+JOB_URLS = json.loads(WORKFLOW_JOBS_STR)
+JOB_URLS = { j['name']: j['url'] for j in JOB_URLS }
+
+REGIONS_BY_SOC = defaultdict(set)     # { soc: set(regions) }
+
+# Import and process data for each board
+for board_data in ALL_BOARD_DATA.values():
+    # Extract common fields from board data
+    artifact = board_data['artifact']
+    board = board_data['board']
+    variant = board_data['variant']
+    subarch = board_data['subarch']
+
+    # Get job link for this build
+    # NOTE: This must match the name used in the workflow YAML
+    job_link = JOB_URLS.get(f"Build for {board}")
+    if job_link:
+        job_link += "#step:5:1"
+
+    # Read loader build data
+    BOARD_LOADERS[board] = LoaderEntry(artifact, board, variant, job_link)
+    if BOARD_LOADERS[board].status == FAILURE:
+        log_test(artifact, board, 'CI test', '', [], FAILURE, "Core data could not be read.")
+        continue
+
+    # Get SoC-specific info from loader config
+    soc = BOARD_LOADERS[board].config['CONFIG_SOC']
+    board_data['soc'] = soc
+    REGIONS_BY_SOC[soc].update(BOARD_LOADERS[board].meminfo.keys())
+
+    # Get list of expected errors for this board
+    exceptions = []
+    if os.path.exists(f"variants/{variant}/known_example_issues.txt"):
+        with open(f"variants/{variant}/known_example_issues.txt", 'r') as f:
+            for line in f:
+                sketch_pattern = line.split('#')[0].strip()
+                if sketch_pattern:
+                    exceptions.append(re.compile(f"^(ArduinoCore-zephyr/)?{sketch_pattern}"))
+
+    # Get raw data from report files for both static and dynamic linking modes
+    for link_mode in ("static", "dynamic"):
+        # Get job link for this test
+        # NOTE: This must match the name used in the workflow YAML
+        job_link = JOB_URLS.get(f"Test {board}:{link_mode}")
+        if job_link:
+            job_link += "#step:6:1"
+
+        # Extract data from the report file
+        report_file = f"arduino-{subarch}-{board}-link_mode={link_mode}.json"
+        if not os.path.exists(report_file):
+            log_test(artifact, board, 'CI test', '', [], FAILURE, f"Report file for {link_mode} not found.", job_link)
+            continue # Skip to the next board
+
+        try:
+            with open(report_file, 'r') as f:
+                report_data = json.load(f)
+        except Exception as e:
+            log_test(artifact, board, 'CI test', '', [], FAILURE, f"Error reading {link_mode} report file: {e}", job_link)
+            continue # Skip to the next board
+
+        reports = report_data.get('boards', [{}])[0].get('sketches', [])
+        if not reports:
+            log_test(artifact, board, 'CI test', '', [], FAILURE, "Test report for {link_mode} is empty, check CI log.", job_link)
+            continue # Skip to the next board
+
+        # Iterate through individual sketch reports
+        for report in reports:
+            sketch = report.get('name', 'unknown_sketch')
+            success = report.get('compilation_success', False)
+            issues = report.get('issues', [])
+
+            # Replace long absolute paths with '...' for brevity.
+            sketch_issues = [ re.sub(r'(/.+?)((/[^/]+){3}):', r'...\2:', issue) for issue in issues ]
+
+            if not success:
+                status = ERROR
+            elif len(sketch_issues): # Implies warnings/non-critical issues
+                status = WARNING
+            else:
+                status = PASS
+
+            # Log the test result to all data structures
+            log_test(artifact, board, sketch, link_mode, exceptions, status, sketch_issues, job_link)
+
+ARTIFACTS = sorted(ARTIFACT_TESTS.keys())
+
+ci_run_status = max(res.status for res in ARTIFACT_TESTS.values())
+ci_run_passed = ci_run_status in (PASS, WARNING, EXPECTED_ERROR)
+
+# Serialize all data to JSON
+output = {
+    'job_link_stem': JOB_LINK_STEM,
+    'artifacts': ARTIFACTS,
+    'ci_run_passed': ci_run_passed,
+    'all_board_data': ALL_BOARD_DATA,
+    'regions_by_soc': {soc: list(regions) for soc, regions in REGIONS_BY_SOC.items()},
+    'board_loaders': {board: loader.to_dict() for board, loader in BOARD_LOADERS.items()},
+    'all_tests': [t.to_dict() for t in ALL_TESTS],
+}
+
+with open(json_file, 'w') as f:
+    json.dump(output, f)
