+++ extra/ci_generate_report.py
@@ -3,232 +3,58 @@
 # Copyright (c) Arduino s.r.l. and/or its affiliated companies
 # SPDX-License-Identifier: Apache-2.0
 
-# Script to analyze CI build and test logs, summarize results,
-# and generate a markdown report.
+# Script to generate CI markdown reports from a JSON data file.
 #
-# The script reads build artifacts and test reports generated during a GitHub
-# Actions CI run for Arduino Zephyr cores. It aggregates results by board,
-# artifact, and sketch, reports failures and warnings, and produces a
-# comprehensive Github-friendly Markdown report.
-#
-# This scripts expects the following inputs:
-#
-#  - Environment variables:
-#     - ALL_BOARD_DATA: JSON array of all board metadata, generated by get_board_details.sh
-#     - WORKFLOW_JOBS: JSON array of job names and URLs, generated by workflow YAML
-#     - GITHUB_REPOSITORY, GITHUB_RUN_ID: workflow context variables for writing links
-#
-#  - Files in the current directory:
-#     - zephyr-<variant>.config: Kconfig settings, generated at loader build time
-#     - zephyr-<variant>.warnings: Loader compilation warnings
-#     - zephyr-<variant>.meminfo: Loader memory usage report
-#     - arduino-<subarch>-<board>-link_mode=<mode>.json: Compilation test report files
-#     - variants/<variant>/known_example_issues.txt: Expected errors for that variant
-#
-# This script produces three output files:
-#
-#  - <results_file>: simple text file with "PASSED" in it if all tests passed, or an error message
+# This script reads the JSON file produced by ci_collect_data.py and generates
+# the same three output files as the original ci_inspect_logs.py script:
+#  - <results_file>: simple text file with "PASSED" or an error message
 #  - <summary_file>: markdown summary report with overall results
 #  - <full_report_file>: full markdown report with detailed logs
 #
+# Usage: ci_generate_report.py <json_file> [<results_file> <summary_file> <full_report_file>]
+#  If output files are not specified, summary and full report are written to /dev/stdout
+#  and results to /dev/null.
 
 from collections import defaultdict
 import json
-import os
-import re
 import sys
 
-# Constants for test statuses, in increasing order of severity
-SKIP = -1           # Test was not performed
-PASS = 0            # (PASS)  Compiled successfully
-WARNING = 1         # (PASS)  Compiled with warnings
-EXPECTED_ERROR = 2  # (PASS*) Compilation failed with expected errors
-ERROR = 3           # (FAIL)  Compilation failed with errors
-FAILURE = 4         # Test run failed to complete
+import ci_shared
+from ci_shared import (
+    SKIP, PASS, WARNING, EXPECTED_ERROR, ERROR, FAILURE,
+    TEST_LEGEND, TEST_STATUS, BOARD_STATUS,
+    LoaderEntry, TestEntry, TestGroup,
+)
 
-# Status legends and icons, indexed by status constant
-TEST_LEGEND = [
-        "Test passed successfully, with no warnings or errors.",
-        "Test completed with some warnings; no errors detected.",
-        "Test completed with errors, but all are known/expected.",
-        "Test completed with unexpected errors.",
-        "Test run failed to complete.",
-        "Test was skipped." # -1
-]
+BASE_COLOR = 0x20               # darkest tone
+DELTA_COLOR = 0xff-2*BASE_COLOR # max color change (brightest tone - darkest tone)
 
-TEST_STATUS = [
-    ":green_circle:",
-    ":yellow_circle:",
-    ":no_entry_sign:",
-    ":red_circle:",
-    ":fire:",
-    ":new_moon:" # -1
-]
+def get_percent(values):
+    if not values:
+        return 0.0
+    return values[0] / values[1]
 
-BOARD_STATUS = [
-    ":white_check_mark:",
-    ":white_check_mark:*",
-    ":heavy_check_mark:*",
-    ":x:",
-    ":fire:",
-    ":new_moon:" # -1
-]
-
-# Loader build status data structure, one per loader
-class LoaderEntry:
-    def __init__(self, artifact, board, variant, job_link):
-        self.artifact = artifact
-        self.board = board
-        self.variant = variant
-        self.job_link = job_link
-
-        self.warnings = self.read_warnings() # list of warning messages
-        self.config = self.read_config() # set of Kconfig settings
-        self.meminfo = self.read_meminfo() # memory usage report
-
-        if not (self.config and self.meminfo):
-            self.status = FAILURE
-        elif self.warnings:
-            self.status = WARNING
-        else:
-            self.status = PASS
-
-    def read_config(self):
-        """
-        Extract Kconfig settings from the loader config file.
-        Format: SYMBOL=VALUE per line or # comment lines for disabled symbols.
-        """
-        report_file = f"zephyr-{self.variant}.config"
-        try:
-            with open(report_file, 'r') as f:
-                configs = {}
-                for line in f:
-                    if line.startswith('#') or '=' not in line:
-                        continue
-                    sym, val = line.strip().split('=', 1)
-                    if val.startswith('"'):
-                        configs[sym] = val.strip('"')
-                    elif val=='y':
-                        configs[sym] = 1
-                    else:
-                        configs[sym] = eval(val)
-                return configs
-        except Exception as e:
-            return {}
-
-    def read_warnings(self):
-        """
-        Read loader compilation warnings from the warnings file.
-        File may be missing in case of no warnings.
-        Format: one line per warning message.
-        """
-        report_file = f"zephyr-{self.variant}.warnings"
-        try:
-            with open(report_file, 'r') as f:
-                return [ line.strip() for line in f if line.strip() ]
-        except Exception as e:
-            return []
-
-    def read_meminfo(self):
-        """
-        Extract memory usage data from the meminfo file.
-        Input format: JSON array of region_name, used_bytes, total_bytes tuples
-        Output format: dict of { region_name: [ used_bytes, total_bytes ] }
-        """
-        report_file = f"zephyr-{self.variant}.meminfo"
-        try:
-            with open(report_file, 'r') as f:
-                report_data = json.load(f)
-                meminfo = { region.replace(':',''): [used, total] for region, used, total in report_data }
-                meminfo.pop('IDT_LIST', None)
-                return meminfo
-        except Exception as e:
-            return {}
-
-# Single test data structure, one per compilation test
-class TestEntry:
-    def __init__(self, artifact, board, sketch, link_mode, excepted, status, issues, job_link):
-        self.artifact = artifact
-        self.board = board
-        self.sketch = sketch
-        self.link_mode = link_mode
-        self.excepted = excepted
-        self.status = EXPECTED_ERROR if excepted and status == ERROR else status
-        self.issues = issues
-        self.job_link = job_link
-        self.invalid_exception = excepted and status in (PASS, WARNING)
-
-        # Extract group and name from sketch path:
-        #                   (1.................) (2....) (3................) (4.)
-        match = re.search(r'(libraries|examples)/([^/]+)/(examples/|extras/)?(.*)', sketch)
-        if match:
-            self.group = match.group(2)
-            self.name = match.group(4)
-        else:
-            self.group = ""
-            self.name = sketch
-
-# Summary data structure, used to group test results
-class TestGroup:
-    def __init__(self):
-        # Sets to track unique board, sketch and (group,name) tuples
-        self.boards = set()
-        self.sketches = set()
-        self.group_names = set()
-        # Counts of test results by status
-        self.counts = { status : 0 for status in [PASS, WARNING, EXPECTED_ERROR, ERROR, FAILURE] }
-        # Overall status of the group
-        self.status = SKIP
-        # List of individual TestEntry objects by feature
-        self.tests = []
-        self.tests_with_issues = []
-        self.tests_with_invalid_exceptions = []
-
-    def track(self, test_entry):
-        """
-        Update this group with a new test entry
-        """
-        self.tests.append(test_entry)
-        if test_entry.issues:
-            self.tests_with_issues.append(test_entry)
-        if test_entry.invalid_exception:
-            self.tests_with_invalid_exceptions.append(test_entry)
-
-        self.counts[test_entry.status] += 1
-        self.status = max(self.status, test_entry.status)
-        self.boards.add(test_entry.board)
-        self.sketches.add(test_entry.sketch)
-        self.group_names.add((test_entry.group, test_entry.name))
-
-# Global Data Structures
-# ----------------------
-
-# Loader build results, one per board
-BOARD_LOADERS = {}                                          # { board: LoaderEntry() }
-
-# Test results grouped by artifact, board, and artifact/sketch
-# (grouping sketch results across different artifacts is really confusing)
-ARTIFACT_TESTS = defaultdict(TestGroup)                     # { artifact: TestGroup() }
-BOARD_TESTS = defaultdict(TestGroup)                        # { board: TestGroup() }
-SKETCH_TESTS = defaultdict(lambda: defaultdict(TestGroup))  # { artifact: { sketch: TestGroup() } }
-
-def log_test(artifact, board, sketch, link_mode, exceptions, status, issues, job_link=None):
+def color_cmd(percent):
     """
-    Logs individual test results into the global test tracking structures.
+    Returns a LaTeX color command string based on the given percentage.
+    The color hue transitions linearly from fully green (0) to fully red (1).
     """
+    color_amt = int(DELTA_COLOR * percent)
+    return f"\\color{{#{BASE_COLOR + color_amt:02x}{0xff - color_amt:02x}{BASE_COLOR:02x}}}"
 
-    # Ensure issues is a list
-    if isinstance(issues, str):
-        issues = [ issues ]
+def color_entry(values, full=True):
+    """
+    Returns a colored LaTeX-formatted string representing memory usage.
+    If full is True, shows used/total as a fraction and percent; otherwise shows only percent.
+    """
+    if not values:
+        return ""
 
-    # Create the test entry
-    excepted = any(pattern.match(sketch) for pattern in exceptions)
-    test_entry = TestEntry(artifact, board, sketch, link_mode, excepted, status, issues, job_link)
-
-    # Track in global structures
-    ARTIFACT_TESTS[artifact].track(test_entry)
-    BOARD_TESTS[board].track(test_entry)
-    SKETCH_TESTS[artifact][sketch].track(test_entry)
+    percent = get_percent(values)
+    if full:
+        return f"${{{color_cmd(percent)}\\frac{{{values[0]}}}{{{values[1]}}}\\space({percent*100:0.1f}\\\\%)}}$"
+    else:
+        return f"{'' if percent < 0.85 else ':warning:'} ${{{color_cmd(percent)}{percent*100:0.1f}\\\\%}}$"
 
 def print_summary():
     """
@@ -433,38 +259,6 @@ def print_test_matrix(artifact, artifact_boards, title, sketch_filter=lambda x: 
 
             f_print("</table></blockquote></details>\n")
 
-REGIONS_BY_SOC = defaultdict(set)     # { soc: set(regions) }
-
-BASE_COLOR = 0x20               # darkest tone
-DELTA_COLOR = 0xff-2*BASE_COLOR # max color change (brightest tone - darkest tone)
-
-def get_percent(values):
-    if not values:
-        return 0.0
-    return values[0] / values[1]
-
-def color_cmd(percent):
-    """
-    Returns a LaTeX color command string based on the given percentage.
-    The color hue transitions linearly from fully green (0) to fully red (1).
-    """
-    color_amt = int(DELTA_COLOR * percent)
-    return f"\\color{{#{BASE_COLOR + color_amt:02x}{0xff - color_amt:02x}{BASE_COLOR:02x}}}"
-
-def color_entry(values, full=True):
-    """
-    Returns a colored LaTeX-formatted string representing memory usage.
-    If full is True, shows used/total as a fraction and percent; otherwise shows only percent.
-    """
-    if not values:
-        return ""
-
-    percent = get_percent(values)
-    if full:
-        return f"${{{color_cmd(percent)}\\frac{{{values[0]}}}{{{values[1]}}}\\space({percent*100:0.1f}\\\\%)}}$"
-    else:
-        return f"{'' if percent < 0.85 else ':warning:'} ${{{color_cmd(percent)}{percent*100:0.1f}\\\\%}}$"
-
 def print_mem_report(artifact, artifact_boards):
     """
     Prints the memory usage report for the given artifact and its boards.
@@ -549,130 +343,52 @@ def print_mem_report(artifact, artifact_boards):
     if extra_data_present:
         f_print("</blockquote></details>")
 
-
-
 # ---------------------------------------------------------------------------
 # Main Logic
 # ---------------------------------------------------------------------------
 
-# Environment Variable Checks
-ALL_BOARD_DATA_STR = os.environ.get('ALL_BOARD_DATA')
-WORKFLOW_JOBS_STR = os.environ.get('WORKFLOW_JOBS')
-GITHUB_REPOSITORY = os.environ.get('GITHUB_REPOSITORY')
-GITHUB_RUN_ID = os.environ.get('GITHUB_RUN_ID')
-JOB_LINK_STEM = f"https://github.com/{GITHUB_REPOSITORY}/actions/runs/{GITHUB_RUN_ID}"
-
-if not ALL_BOARD_DATA_STR or not GITHUB_REPOSITORY or not GITHUB_RUN_ID:
-    print("Not in a Github CI run, cannot proceed.")
-    sys.exit(0)
-
-if not len(sys.argv) in (1, 4):
-    print("Usage: ci_inspect_logs.py [<results_file> <summary_file> <full_report_file>]")
+if not len(sys.argv) in (2, 5):
+    print("Usage: ci_generate_report.py <json_file> [<results_file> <summary_file> <full_report_file>]")
     sys.exit(1)
 
-if len(sys.argv) == 4:
-    results_file = sys.argv[1]
-    summary_file = sys.argv[2]
-    full_report_file = sys.argv[3]
+json_file = sys.argv[1]
+
+if len(sys.argv) == 5:
+    results_file = sys.argv[2]
+    summary_file = sys.argv[3]
+    full_report_file = sys.argv[4]
 else:
     results_file = "/dev/null"
     summary_file = full_report_file = "/dev/stdout"
 
-# Load board data and job URLs from environment variables
-ALL_BOARD_DATA = json.loads(ALL_BOARD_DATA_STR)
-ALL_BOARD_DATA = { b['board']: b for b in ALL_BOARD_DATA }
+# Load data from JSON
+with open(json_file, 'r') as f:
+    data = json.load(f)
 
-JOB_URLS = json.loads(WORKFLOW_JOBS_STR)
-JOB_URLS = { j['name']: j['url'] for j in JOB_URLS }
+JOB_LINK_STEM = data['job_link_stem']
+ARTIFACTS = data['artifacts']
+ci_run_passed = data['ci_run_passed']
+ALL_BOARD_DATA = data['all_board_data']
+REGIONS_BY_SOC = {soc: set(regions) for soc, regions in data['regions_by_soc'].items()}
 
-# Import and process data for each board
-for board_data in ALL_BOARD_DATA.values():
-    # Extract common fields from board data
-    artifact = board_data['artifact']
-    board = board_data['board']
-    variant = board_data['variant']
-    subarch = board_data['subarch']
+# Reconstruct BOARD_LOADERS from JSON
+BOARD_LOADERS = {
+    board: LoaderEntry.from_dict(loader)
+    for board, loader in data['board_loaders'].items()
+}
 
-    # Get job link for this build
-    # NOTE: This must match the name used in the workflow YAML
-    job_link = JOB_URLS.get(f"Build for {board}")
-    if job_link:
-        job_link += "#step:5:1"
+# Reconstruct all test entries and rebuild groups via track()
+ALL_TESTS = [TestEntry.from_dict(t) for t in data['all_tests']]
 
-    # Read loader build data
-    BOARD_LOADERS[board] = LoaderEntry(artifact, board, variant, job_link)
-    if BOARD_LOADERS[board].status == FAILURE:
-        log_test(artifact, board, 'CI test', '', [], FAILURE, "Core data could not be read.")
-        continue
+ARTIFACT_TESTS = defaultdict(TestGroup)
+BOARD_TESTS = defaultdict(TestGroup)
+SKETCH_TESTS = defaultdict(lambda: defaultdict(TestGroup))
+for t in ALL_TESTS:
+    ARTIFACT_TESTS[t.artifact].track(t)
+    BOARD_TESTS[t.board].track(t)
+    SKETCH_TESTS[t.artifact][t.sketch].track(t)
 
-    # Get SoC-specific info from loader config
-    soc = BOARD_LOADERS[board].config['CONFIG_SOC']
-    board_data['soc'] = soc
-    REGIONS_BY_SOC[soc].update(BOARD_LOADERS[board].meminfo.keys())
-
-    # Get list of expected errors for this board
-    exceptions = []
-    if os.path.exists(f"variants/{variant}/known_example_issues.txt"):
-        with open(f"variants/{variant}/known_example_issues.txt", 'r') as f:
-            for line in f:
-                sketch_pattern = line.split('#')[0].strip()
-                if sketch_pattern:
-                    exceptions.append(re.compile(f"^(ArduinoCore-zephyr/)?{sketch_pattern}"))
-
-    # Get raw data from report files for both static and dynamic linking modes
-    for link_mode in ("static", "dynamic"):
-        # Get job link for this test
-        # NOTE: This must match the name used in the workflow YAML
-        job_link = JOB_URLS.get(f"Test {board}:{link_mode}")
-        if job_link:
-            job_link += "#step:6:1"
-
-        # Extract data from the report file
-        report_file = f"arduino-{subarch}-{board}-link_mode={link_mode}.json"
-        if not os.path.exists(report_file):
-            log_test(artifact, board, 'CI test', '', [], FAILURE, f"Report file for {link_mode} not found.", job_link)
-            continue # Skip to the next board
-
-        try:
-            with open(report_file, 'r') as f:
-                report_data = json.load(f)
-        except Exception as e:
-            log_test(artifact, board, 'CI test', '', [], FAILURE, f"Error reading {link_mode} report file: {e}", job_link)
-            continue # Skip to the next board
-
-        reports = report_data.get('boards', [{}])[0].get('sketches', [])
-        if not reports:
-            log_test(artifact, board, 'CI test', '', [], FAILURE, "Test report for {link_mode} is empty, check CI log.", job_link)
-            continue # Skip to the next board
-
-        # Iterate through individual sketch reports
-        for report in reports:
-            sketch = report.get('name', 'unknown_sketch')
-            success = report.get('compilation_success', False)
-            issues = report.get('issues', [])
-
-            # Replace long absolute paths with '...' for brevity.
-            sketch_issues = [ re.sub(r'(/.+?)((/[^/]+){3}):', r'...\2:', issue) for issue in issues ]
-
-            if not success:
-                status = ERROR
-            elif len(sketch_issues): # Implies warnings/non-critical issues
-                status = WARNING
-            else:
-                status = PASS
-
-            # Log the test result to all data structures
-            log_test(artifact, board, sketch, link_mode, exceptions, status, sketch_issues, job_link)
-
-ARTIFACTS = sorted(ARTIFACT_TESTS.keys())
-
-# Begin output of the report
-# --------------------------
-
-ci_run_status = max(res.status for res in ARTIFACT_TESTS.values())
-ci_run_passed = ci_run_status in (PASS, WARNING, EXPECTED_ERROR)
-
-# update the results file
+# Update the results file
 with open(results_file, 'w') as f:
     f.write('Tests PASSED.' if ci_run_passed else '::error::Tests FAILED.');
     f.write(f" See {JOB_LINK_STEM}#user-content-summary for details.\n")
